requirements.txt
asgiref==3.5.2
Django==4.0.6
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
Unidecode==1.3.4
django-filter==23.4
Pillow==10.1.0
django-jazzmin==2.6.0
drf-yasg==1.21.7
requests
djangorestframework-simplejwt
psycopg2